try to refacto AdminPage to single view and component

+++ backend/controllers/userController.js
@@ -190,6 +190,12 @@ export const deleteUser = async (req, res) => {
       return res.status(404).json({ error: "Utilisateur non trouvé" });
     }
 
+    if (user.isAdmin === true) {
+      return res
+        .status(401)
+        .json({ message: "Vous ne pouvez pas supprimer un Administrateur" });
+    }
+
     // Si la clé secrète est égale à "process.env.SECRET_ADMIN_KEY", supprime l'utilisateur sans vérifier le mot de passe
     if (secretadminkey === process.env.SECRET_ADMIN_KEY) {
       await User.findByIdAndRemove(id);
